refactor: split panel_test_fixtures into focused modules (#3847)
Replace the 1171-line monolith with panel_fixtures/* modules behind a thin
barrel export; satisfies AC4 module-size cap with no import-path changes.
